pyproject.toml
[tool.poetry]
name = "pre-commit-hooks"
version = "0.0.1"
description = ""
authors = ["Trim21 <i@trim21.me>"]
packages = [{ include = "pre_commit_hooks" }]

[tool.poetry.dependencies]
python = "^3.6.1"
"ruamel.yaml" = "0.16.10"

[tool.poetry.scripts]
yamlfmt = 'pre_commit_hooks.yamlfmt:main'

[tool.poetry.dev-dependencies]
# tests
pytest = "==5.4.3"
coverage = "^5.1"
# linter and formatter
pre-commit = "==2.4.0"
flake8 = '==3.8.2'
flake8-comprehensions = "3.2.2"
flake8-bugbear = "20.1.4"
flake8-commas = "2.0.0"
flake8-mutable = "1.2.0"
flake8-broken-line = "0.2.0"
flake8-pytest-style = "1.1.1"
mypy = "==0.780"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
